Igor

+++ app/src/main/java/by/isb/weatherwidget/data/dto/forecast/ForecastResponse.kt
@@ -39,9 +39,9 @@ data class ForecastResponse(
         @SerializedName("sunset")
         val sunset: Int?,
         @SerializedName("temp")
-        val temp: Int?,
+        val temp: Double?,
         @SerializedName("uvi")
-        val uvi: Int?,
+        val uvi: Double?,
         @SerializedName("visibility")
         val visibility: Int?,
         @SerializedName("weather")
@@ -71,7 +71,7 @@ data class ForecastResponse(
         @SerializedName("dew_point")
         val dewPoint: Double?,
         @SerializedName("dt")
-        val dt: Int?,
+        val dt: Long?,
         @SerializedName("feels_like")
         val feelsLike: FeelsLike?,
         @SerializedName("humidity")
@@ -155,15 +155,15 @@ data class ForecastResponse(
         @SerializedName("humidity")
         val humidity: Int?,
         @SerializedName("pop")
-        val pop: Int?,
+        val pop: Double?,
         @SerializedName("pressure")
         val pressure: Int?,
         @SerializedName("rain")
         val rain: Rain?,
         @SerializedName("temp")
-        val temp: Int?,
+        val temp: Double?,
         @SerializedName("uvi")
-        val uvi: Int?,
+        val uvi: Double?,
         @SerializedName("visibility")
         val visibility: Int?,
         @SerializedName("weather")
